requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
faiss-cpu
numpy==1.24.3
openai
PyPDF2==3.0.1
pydantic==2.5.0